Version 2 du projet avec un nouveau modal
